requirements.txt
aiosqlite==0.22.1
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
click==8.1.8
cryptography==46.0.5
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.124.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
sqlmodel==0.0.31
starlette==0.49.3
tomli==2.3.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
